simulator/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
databases==0.9.0
exceptiongroup==1.2.2
fastapi==0.112.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
kombu==5.4.0
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
